Past orders page with all functionalities
This is the past orders page for the user. I have implemented all functionalities such as the profile button, which will make the Guthaben button pop up so that the user can be directed to that page. When clicking on the LS Logo on the top left, the user will be directed to the main page. The more-button on the top right hand side also works fully, but as the user is already on the past orders page, clicking on the ,,vergangene Bestellungen'' button will refresh the site.

Co-Authored-By: Mohamed Amin Ben Chrifa <mohamed6aminbenchrifa@gmail.com>