The cart- and search pages have following changes
The bell icon works dynamically and updates every 5 seconds for changes. I added a console log to ensure it works properly. Furthermore I added the ausloggen button which also directs to the index page. I separated the search bar on the search page into two parts, where restaurant name and cuisine need to be typed in. I will now test with dummys if it works as expected.