checkout page works
The popup now shows the sum of all items. If there is no item in the cart, it will display it accordingly. After purchase, the user will be directed to the search page. Btw. as soon as I finish the cart page, I will work on minor bugs on the page like the error message.